Compiles after some static_casts, IntGen and ProtoIntGen might be better fixed..

+++ src/Generating/IntGen.h
@@ -1,4 +1,4 @@
-
+#include <iostream>
 // IntGen.h
 
 // Declares the cIntGen class and descendants for generating and filtering various 2D arrays of ints
@@ -494,7 +494,8 @@ public:
 					if (IsBiomeOcean(above) || IsBiomeOcean(below) || IsBiomeOcean(left) || IsBiomeOcean(right))
 					{
 						// First convert the value to a regular biome (drop the M flag), then modulo by our biome count:
-						val = ToBeach[(val % 128) % ARRAYCOUNT(ToBeach)];
+						// KINGCOL13TODO: work out if this can all be ints
+						val = ToBeach[static_cast<unsigned long>((val % 128)) % ARRAYCOUNT(ToBeach)];
 					}
 				}
 				a_Values[x + z * SizeX] = val;
